🚚 rename functions for clarity devConfig -> generateDevConfig ect
